Significant restructuring of pursuit learner
