+++ adam/experiment/log_experiment.py
@@ -20,7 +20,8 @@ from adam.experiment.observer import LearningProgressHtmlLogger, CandidateAccura
 from adam.learner import LanguageLearner
 from adam.learner.object_recognizer import ObjectRecognizer
 from adam.learner.preposition_subset import PrepositionSubsetLanguageLearner
-from adam.learner.pursuit import ObjectPursuitLearner, HypothesisLogger
+from adam.learner.pursuit import HypothesisLogger
+from adam.learner.pursuit.object import ObjectPursuitLearner
 from adam.ontology.phase1_ontology import GAILA_PHASE_1_ONTOLOGY
 from adam.perception.high_level_semantics_situation_to_developmental_primitive_perception import (
     GAILA_M6_PERCEPTION_GENERATOR,
